+++ common/src/cpx_op.c
@@ -1,18 +1,18 @@
 #include "cpx_op.h"
 
-long double* tms(long double* trgt, long double* src1, long double* src2){
+double* tms(double* trgt, double* src1, double* src2){
   trgt[0]=src1[0]*src2[0]-src1[1]*src2[1];
   trgt[1]=src1[1]*src2[0]+src1[0]*src2[1];
   return trgt;
 }
 
-long double* add(long double* trgt, long double* src1, long double* src2){
+double* add(double* trgt, double* src1, double* src2){
   trgt[0]=src1[0]+src2[0];
   trgt[1]=src1[1]+src2[1];
   return trgt;
 }
 
-long double* asn(long double* trgt, long double* src){
+double* asn(double* trgt, double* src){
   trgt[0]=src[0]; trgt[1]=src[1];
   return trgt;
 }
